requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
Requests==2.32.4
schedule==1.2.2
